# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: StellarRoute API
  version: 0.1.0
  description: |
    REST API for DEX aggregation on the Stellar Network.

    StellarRoute aggregates liquidity from Stellar's SDEX (Stellar Distributed
    Exchange) and Soroban AMM pools to find the best available price for any
    token pair.

    ## Asset identifiers

    Assets are encoded in path segments using one of three formats:

    | Format | Example | Meaning |
    |--------|---------|---------|
    | `native` | `native` | XLM (Stellar's native asset) |
    | `CODE` | `USDC` | Issued asset — code only (omits issuer) |
    | `CODE:ISSUER` | `USDC:GA5ZSEJ...` | Issued asset — fully qualified |

    ## Error handling

    All errors return a JSON body with `error` (machine-readable code) and
    `message` (human-readable description).  See the `ErrorResponse` schema
    and the **Error codes** section below for the full list.

    ## Rate limiting

    Endpoints are rate-limited per IP address.  When the limit is exceeded the
    API returns **429 Too Many Requests** with `X-RateLimit-*` headers.

    ## Request correlation

    Clients may optionally send `X-Request-ID` with any request. The API will
    reuse that value, or generate a UUID when the header is omitted, and echo
    the final request ID back in the `X-Request-ID` response header.

    ## Versioning and deprecation

    `/api/v1/*` is in deprecation mode. Responses from v1 routes include
...
    notice period and migration guidance.
...
servers:
...
paths:
...
  /api/v1/assets:
    get:
      operationId: list_assets_metadata
      summary: Bulk asset metadata
      description: |
        Returns metadata for multiple assets identified by their codes.
      tags:
        - assets
      parameters:
        - name: codes
          in: query
          description: Comma-separated list of asset codes (e.g. `XLM,USDC`)
          required: true
          schema:
            type: string
            example: XLM,USDC
      responses:
        "200":
          description: List of asset metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetMetadataBulkResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/assets/{code}:
    get:
      operationId: get_asset_metadata
      summary: Get asset metadata
      description: |
        Returns decimals, issuer labels, and domain information for a single asset.
      tags:
        - assets
      parameters:
        - name: code
          in: path
          description: Asset code (e.g. `XLM`, `USDC`)
          required: true
          schema:
            type: string
            example: USDC
        - name: issuer
          in: query
          description: Optional issuer address for disambiguating common codes.
          required: false
          schema:
            type: string
            example: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
      responses:
        "200":
          description: Asset metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetMetadataResponse"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/pairs:
    get:
      operationId: list_pairs
      summary: List trading pairs
      description: |
        Returns active trading pairs that currently have offers in the SDEX
        orderbook.

        Stable sort order:
        1. `offer_count DESC`
        2. `last_updated DESC`
        3. Canonical asset fields ascending as deterministic tie-breakers

        Pagination:
        - Default page size: `25`
        - Max page size: `100`
        - `cursor` (preferred) or `offset` (fallback)
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOffset"
      responses:
        "200":
          description: List of active trading pairs
          headers:
            Deprecation:
              $ref: "#/components/headers/DeprecationHeader"
            Sunset:
              $ref: "#/components/headers/SunsetHeader"
            Link:
              $ref: "#/components/headers/DeprecationLinkHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairsResponse"
              example:
                pairs:
                  - base: XLM
                    counter: USDC
                    base_asset: native
                    counter_asset: "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"
                    offer_count: 42
                    last_updated: "2026-02-23T11:59:00Z"
                  - base: USDC
                    counter: BTC
                    base_asset: "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"
                    counter_asset: "BTC:GAUTUYY2THLF7SGITDFMXJVYH3LHDSMGEAKSBU267M2K7A3W543CKUEF"
                    offer_count: 17
                    last_updated: "2026-02-23T11:58:30Z"
                total: 2
                limit: 25
                next_cursor: "25"
                prev_cursor: null
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: validation_error
                  message: Invalid cursor; expected a numeric offset
        "404":
          description: Trading pairs not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: not_found
                  message: No trading pairs found
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: internal_error
                  message: An internal error occurred
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/markets:
    get:
      operationId: list_markets
      summary: List markets
      description: Alias of `/api/v1/pairs` with identical pagination and sorting.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOffset"
      responses:
        "200":
          description: List of active markets
          headers:
            Deprecation:
              $ref: "#/components/headers/DeprecationHeader"
            Sunset:
              $ref: "#/components/headers/SunsetHeader"
            Link:
              $ref: "#/components/headers/DeprecationLinkHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairsResponse"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/orderbook/{base}/{quote}:
    get:
      operationId: get_orderbook
      summary: Get orderbook
      description: |
        Returns bids and asks for the given base/quote asset pair.

        **Bids** are orders to buy the base asset with the quote asset
        (sorted highest price first).  **Asks** are orders to sell the base
        asset for the quote asset (sorted lowest price first).

        Results are cached for **10 seconds**.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/BaseAsset"
        - $ref: "#/components/parameters/QuoteAsset"
      responses:
        "200":
          description: Orderbook data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderbookResponse"
              example:
                base_asset:
                  asset_type: native
                  asset_code: null
                  asset_issuer: null
                quote_asset:
                  asset_type: credit_alphanum4
                  asset_code: USDC
                  asset_issuer: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
                bids:
                  - price: "0.1050000"
                    amount: "500.0000000"
                    total: "52.5000000"
                asks:
                  - price: "0.1060000"
                    amount: "300.0000000"
                    total: "31.8000000"
                summary:
                  bid: "0.1050000"
                  ask: "0.1060000"
                  spread_bps: 95
                  midpoint: "0.1055000"
                timestamp: 1740312000
        "400":
          description: Invalid asset identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: invalid_asset
                  message: "Invalid base asset: unknown asset format"
        "404":
          description: Asset pair not found in the orderbook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: not_found
                  message: Asset not found in orderbook
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: internal_error
                  message: An internal error occurred
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/price-history/{base}/{quote}:
    get:
      operationId: get_price_history
      summary: Get 24h price history
      description: |
        Returns a lightweight 24h sparkline series for the selected trading pair.

        The backend derives the series from `orderbook_snapshots.mid_price`
        using hourly aggregation buckets. If no snapshots are available in the
        trailing 24h window, the endpoint returns an empty `points` array.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/BaseAsset"
        - $ref: "#/components/parameters/QuoteAsset"
      responses:
        "200":
          description: 24h price history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistoryResponse"
              example:
                base_asset:
                  asset_type: native
                  asset_code: null
                  asset_issuer: null
                quote_asset:
                  asset_type: credit_alphanum4
                  asset_code: USDC
                  asset_issuer: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
                window: "24h"
                source: orderbook_snapshots.mid_price
                generated_at: 1740312000000
                points:
                  - timestamp: 1740225600000
                    price: "0.1051000"
                  - timestamp: 1740229200000
                    price: "0.1054000"
        "400":
          description: Invalid asset identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Trading pair not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/quote/{base}/{quote}:
    get:
      operationId: get_quote
      summary: Get price quote
      description: |
        Returns the best available price and routing path for trading `amount`
        of the base asset for the quote asset (or vice-versa when
        `quote_type=buy`).

        The router evaluates both SDEX offers and Soroban AMM pools to find
        the optimal execution path.

        Results are cached for **5 seconds**. Large responses honor
        `Accept-Encoding: br` or `Accept-Encoding: gzip`; smaller responses
        remain uncompressed. Prometheus metrics track original versus wire
        response bytes.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/BaseAsset"
        - $ref: "#/components/parameters/QuoteAsset"
        - $ref: "#/components/parameters/RequestIdHeader"
        - name: amount
          in: query
          description: Amount of the base asset to trade (default `1`)
          required: false
          schema:
            type: string
            example: "100"
        - name: slippage_bps
          in: query
          description: Slippage tolerance in basis points (e.g. `50` = 0.50%). Max `10000` (100%).
          required: false
          schema:
            type: integer
            default: 50
            example: 100
        - name: quote_type
          in: query
          description: |
            Direction of the quote:
            - `sell` — how much quote asset you receive when **selling** `amount` of the base asset
            - `buy`  — how much base asset you must spend to **buy** `amount` of the quote asset
          required: false
          schema:
            type: string
            enum: [sell, buy]
            default: sell
        - name: Accept-Encoding
          in: header
          description: Optional response compression negotiation. Large quote responses support `br` and `gzip`; smaller responses remain uncompressed.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Best available price quote
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
            Vary:
              schema:
                type: string
              description: Includes `Accept-Encoding` when quote response compression is negotiated.
            Content-Encoding:
              schema:
                type: string
                enum: [br, gzip]
              description: Present when the quote response is compressed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
              example:
                base_asset:
                  asset_type: native
                  asset_code: null
                  asset_issuer: null
                quote_asset:
                  asset_type: credit_alphanum4
                  asset_code: USDC
                  asset_issuer: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
                amount: "100.0000000"
                price: "0.1055000"
                total: "10.5500000"
                quote_type: sell
                degraded: false
                path:
                  - from_asset:
                      asset_type: native
                      asset_code: null
                      asset_issuer: null
                    to_asset:
                      asset_type: credit_alphanum4
                      asset_code: USDC
                      asset_issuer: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
                    price: "0.1055000"
                    source: sdex
                timestamp: 1740312000
        "400":
          description: Invalid parameters (bad asset format or non-positive amount)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: validation_error
                  message: Amount must be greater than zero
        "404":
          description: No trading route found for this pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: no_route
                  message: No trading route found for this pair
        "422":
          description: Stale market data — all data sources are too old to provide a quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: stale_market_data
                message: All market data inputs are stale
                details:
                  stale_count: 5
                  fresh_count: 0
                  threshold_secs_sdex: 60
                  threshold_secs_amm: 120
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: internal_error
                  message: An internal error occurred
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/route/{base}/{quote}:
    get:
      operationId: get_route
      summary: Get trading route
      deprecated: true
      description: |
        Returns only the optimal execution path and routing steps for trading
        `amount` of the base asset for the quote asset.

        Unlike the quote endpoint, this returns a simplified response without
        detailed pricing breakdown or venue evaluations.

        This legacy route is deprecated in favor of `/api/v1/routes/{base}/{quote}`.
        Clients should watch the `Deprecation`, `Sunset`, and `Link` headers and
        migrate using the guidance in `docs/api/versioning.md`.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/BaseAsset"
        - $ref: "#/components/parameters/QuoteAsset"
        - name: amount
          in: query
          description: Amount of the base asset to trade (default `1`)
          required: false
          schema:
            type: string
            example: "100"
        - name: slippage_bps
          in: query
          description: Slippage tolerance in basis points (e.g. `50` = 0.50%). Max `10000` (100%).
          required: false
          schema:
            type: integer
            default: 50
            example: 100
        - name: quote_type
          in: query
          description: |
            Direction of the quote:
            - `sell` — how much quote asset you receive when **selling** `amount` of the base asset
            - `buy`  — how much base asset you must spend to **buy** `amount` of the quote asset
          required: false
          schema:
            type: string
            enum: [sell, buy]
            default: sell
      responses:
        "200":
          description: Optimal trading route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "400":
          description: Invalid parameters (bad asset format or invalid slippage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No trading route found for this pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      description: Optional client-supplied correlation ID echoed back by the API.
      schema:
        type: string
        example: support-ticket-42
...
    PageLimit:
      name: limit
      in: query
      required: false
      description: Page size. Default 25, max 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    PageCursor:
      name: cursor
      in: query
      required: false
      description: Cursor for the next page (numeric offset token).
      schema:
        type: string
        example: "25"
    PageOffset:
      name: offset
      in: query
      required: false
      description: Offset fallback for pagination when cursor is not used.
      schema:
        type: integer
        minimum: 0
        example: 50
...
  headers:
...
    DeprecationHeader:
      description: Indicates that the endpoint version is deprecated.
      schema:
        type: string
        example: "true"
    SunsetHeader:
      description: RFC 7231 IMF-fixdate indicating planned v1 sunset.
      schema:
        type: string
        example: "Wed, 31 Dec 2026 23:59:59 GMT"
    DeprecationLinkHeader:
      description: Link to migration guidance for deprecated v1 routes.
      schema:
        type: string
        example: '</docs/api/v1-migration-guide>; rel="deprecation"'
  schemas:
...
    PairsResponse:
      type: object
      required: [pairs, total]
      description: List of active trading pairs
      properties:
        pairs:
          type: array
          items:
            $ref: "#/components/schemas/TradingPair"
        total:
          type: integer
          description: Total number of pairs available across all pages
          example: 2
        limit:
          type: integer
          nullable: true
          description: Effective page size used for this response
          example: 25
        next_cursor:
          type: string
          nullable: true
          description: Cursor token for the next page, if any
          example: "25"
        prev_cursor:
          type: string
          nullable: true
          description: Cursor token for the previous page, if any
          example: "0"
    AssetMetadataResponse:
      type: object
      required: [code, decimals, asset_type]
      description: Enriched asset metadata
      properties:
        code:
          type: string
          description: Asset code
          example: USDC
        issuer:
          type: string
          nullable: true
          description: G-address of the issuer
          example: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
        decimals:
          type: integer
          description: Number of decimal places for the asset
          example: 7
        asset_type:
          type: string
          description: Stellar asset type
          example: credit_alphanum4
        display_name:
          type: string
          nullable: true
          description: Human-readable display name or issuer label
          example: USDC
        icon_url:
          type: string
          nullable: true
          format: uri
          description: URL to the asset's icon/logo
          example: https://stellar.org/images/lumens-logo.svg
        domain:
          type: string
          nullable: true
          description: Home domain of the issuer
          example: centre.io
    AssetMetadataBulkResponse:
      type: object
      required: [assets]
      properties:
        assets:
          type: array
          items:
            $ref: "#/components/schemas/AssetMetadataResponse"
...
    ErrorResponse:
      type: object
      required: [error, message]
      description: Standard API error format
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: validation_error
        message:
          type: string
          description: Human-readable error description
          example: "Amount must be greater than zero"
        details:
          type: object
          nullable: true
          description: Optional structured context about the error
          example:
            parameter: slippage_bps
            min: 0
            max: 10000

  # Authentication is planned for a future release.
  # securitySchemes:
  #   bearerAuth:
  #     type: http
  #     scheme: bearer
  #     bearerFormat: JWT
  #     description: |
  #       API key authentication (Phase 2).
  #       Pass a JWT obtained from the /auth/token endpoint.